wai-aria/heading/invalid-nesting.html

<!doctype html>
<html>

<head>
  <meta charset="utf-8">
  <title>WAI-ARIA tests for invalidly nested elements with heading role</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>

<body>

  <h1>WAI-ARIA tests for invalidly nested heading elements</h1>
  <p>Equivalent ARIA patterns tested in <code>/html-aam/heading/invalid-nesting.html</code>.</p>

  <h2>Accessible names and roles</h2>
  <hr>

  <div
    role="heading"
    aria-level="1"
  >
    <div
      role="heading"
      aria-level="2"
      class="ex-label ex-role"
      data-testname="Invalidly nested div[role=heading][aria-level=2] should have accname matching its text contents, and should still convey heading role"
      data-expectedlabel="Invalid div[role=heading][aria-level=2]"
      data-expectedrole="heading"
    >Invalid div[role=heading][aria-level=2]</div>
  </div>

  <div
    role="heading"
    aria-level="4"
  >
    <div
      role="heading"
      aria-level="3"
      class="ex-label ex-role"
      data-testname="Invalidly nested div[role=heading][aria-level=3] should have accname matching its text contents, and should still convey heading role"
      data-expectedlabel="Invalid div[role=heading][aria-level=3]"
      data-expectedrole="heading"
    >Invalid div[role=heading][aria-level=3]</div>
  </div>

  <div
    role="heading"
    aria-level="5"
  >
    <div
      role="heading"
      aria-level="6"
      class="ex-label ex-role"
      data-testname="Invalidly nested div[role=heading][aria-level=6] should have accname matching its text contents, and should still convey heading role"
      data-expectedlabel="Invalid div[role=heading][aria-level=6]"
      data-expectedrole="heading"
    >Invalid div[role=heading][aria-level=6]</div>
  </div>

  <h2>ARIA heading impostors</h2>
  <hr>

  <button
    role="heading"
    aria-level="1"
  >Macaroni
    <span
      role="heading"
      aria-level="2"
      class="ex-label ex-role"
      data-testname="Invalidly nested span[role=heading][aria-level=2] should have accname matching its text contents, and should still convey heading role"
      data-expectedlabel="Cheese"
      data-expectedrole="heading"
    >Cheese</span>
  </button>

  <script>
    AriaUtils.verifyRolesAndLabelsBySelector(".ex-label.ex-role");
    verifyNoForbiddenAncestorRole(".ex-no-heading-ancestor", "heading");

    function verifyNoForbiddenAncestorRole(selector, forbiddenRole) {
      const els = document.querySelectorAll(selector);
      if (!els.length) {
        throw `Selector passed in verifyNoForbiddenAncestorRole("${selector}") should match at least one element.`;
      }
      for (const el of els) {
        let testName = el.getAttribute("data-testname");
        let ancestors = [];
        let ancestor = el.parentNode;
        promise_test(async t => {
          // Get computed role of each ancestor up to but excluding the root.
          while (ancestor && ancestor != document.documentElement) {
            const computedRole = await test_driver.get_computed_role(ancestor);
            ancestors.push(computedRole);
            ancestor = ancestor.parentNode;
          }
          assert_false(ancestors.includes(forbiddenRole));
        }, `${testName}`);
      }
    }

  </script>
</body>

</html>
